:recycle: Extract test_utils package
